add tensor core gemm

+++ gemm_cuda_core_1.cu
@@ -2,14 +2,14 @@
 
 #include "utils.h"
 
-__global__ void Kernel(const half *A, const half *B, half *C, uint64_t m, uint64_t n, uint64_t k) {
+__global__ void Kernel(const half *A, const half *B, float *C, uint64_t m, uint64_t n, uint64_t k) {
     uint64_t y = blockIdx.y * blockDim.y + threadIdx.y;
     uint64_t x = blockIdx.x * blockDim.x + threadIdx.x;
 
     if (y < m && x < n) {
-        half sum = 0.0f;
+        float sum = 0.0f;
         for (uint64_t t = 0; t < k; ++t) {
-            sum += (A[y * k + t] * B[x * k + t]);
+            sum += __half2float(A[y * k + t] * B[x * k + t]);
         }
         C[y * n + x] = (sum);
     }
@@ -22,7 +22,7 @@ class GemmCudaCore_1 : public GemmBase {
     void LaunchKernel(
         const half *d_A,
         const half *d_B,
-        half *d_C, 
+        float *d_C, 
         uint64_t m, uint64_t n, uint64_t k) override
 {
         dim3 blockdim(16, 16);
