+++ gemm_cuda_core_2.cu
@@ -4,14 +4,14 @@
 
 constexpr uint64_t TILE_SIZE = 16;
 
-__global__ void Kernel(const half *A, const half *B, half *C, uint64_t m, uint64_t n, uint64_t k) {
+__global__ void Kernel(const half *A, const half *B, float *C, uint64_t m, uint64_t n, uint64_t k) {
     __shared__ half s_A[TILE_SIZE][TILE_SIZE];
     __shared__ half s_B[TILE_SIZE][TILE_SIZE];
 
     uint64_t y = blockIdx.y * blockDim.y + threadIdx.y;
     uint64_t x = blockIdx.x * blockDim.x + threadIdx.x;
 
-    half sum = 0.0f;
+    float sum = 0.0f;
     for (uint64_t k_tile = 0; k_tile < k; k_tile += TILE_SIZE) {
         s_A[threadIdx.y][threadIdx.x] = A[y * k + k_tile + threadIdx.x];
         s_B[threadIdx.x][threadIdx.y] = B[x * k + k_tile + threadIdx.y];
@@ -19,7 +19,7 @@ __global__ void Kernel(const half *A, const half *B, half *C, uint64_t m, uint64
 
         if (y < m && x < n) {
             for (uint64_t t = 0; t < TILE_SIZE; ++t) {
-                sum += (s_A[threadIdx.y][t] * s_B[threadIdx.x][t]);
+                sum += __half2float(s_A[threadIdx.y][t] * s_B[threadIdx.x][t]);
             }
         }
         __syncthreads();
@@ -37,7 +37,7 @@ class GemmCudaCore_2 : public GemmBase {
     void LaunchKernel(
         const half *d_A,
         const half *d_B,
-        half *d_C, 
+        float *d_C, 
         uint64_t m, uint64_t n, uint64_t k) override
 {
         dim3 blockdim(TILE_SIZE, TILE_SIZE);
